Fix frontend build issue for Arches installed with pipenv, poetry or uv
Resolves #11763

The convention of naming a virtual environment "`.venv`" appears to have been adopted
by the above package managers.

The fix addresses the issue of templates in directories that contain a dot (specifically
for Arches installed with the above package managers) while ensuring that dotfiles, such
as macOS' .DS_Store (#11094), are ignored.